+++ includes/material.h
@@ -6,7 +6,7 @@
 /*   By: pipolint <pipolint@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/11 21:28:19 by pipolint          #+#    #+#             */
-/*   Updated: 2024/10/11 21:29:05 by pipolint         ###   ########.fr       */
+/*   Updated: 2024/10/22 18:38:32 by pipolint         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,6 @@
 # include "structs.h"
 
 t_mater	*create_default_material();
-t_mater	*create_material(t_color color, float diffuse, float ambient, float specular, float shine);
+t_mater	*create_material(t_color color, double diffuse, double ambient, double specular, double shine);
 
 #endif
