+++ vector_ops.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   vector_ops.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ahaarij <ahaarij@student.42abudhabi.ae>    +#+  +:+       +#+        */
+/*   By: pipolint <pipolint@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/21 16:28:06 by pipolint          #+#    #+#             */
-/*   Updated: 2024/10/22 00:02:26 by ahaarij          ###   ########.fr       */
+/*   Updated: 2024/10/22 18:38:32 by pipolint         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@
 	** subtracting a point(w=1) from a vector(w=0) doesn't make sense
 */
 
-float	dot_product(t_tuple *vec1, t_tuple *vec2)
+double	dot_product(t_tuple *vec1, t_tuple *vec2)
 {
 	return ((vec1->x * vec2->x) + (vec1->y * vec2->y) + (vec1->z * vec2->z));
 }
@@ -30,9 +30,9 @@ float	dot_product(t_tuple *vec1, t_tuple *vec2)
 t_tuple	cross_product(t_tuple *vec1, t_tuple *vec2)
 {
 	t_tuple	cross;
-	float	x;
-	float	y;
-	float	z;
+	double	x;
+	double	y;
+	double	z;
 
 	ft_bzero(&cross, sizeof(t_tuple));
 	x = vec1->y * vec2->z - vec1->z * vec2->y;
@@ -44,7 +44,7 @@ t_tuple	cross_product(t_tuple *vec1, t_tuple *vec2)
 
 void	normalize(t_tuple *vector)
 {
-	float	magnitude;
+	double	magnitude;
 
 	magnitude = sqrt(((vector->x * vector->x) + (vector->y * vector->y) + \
 		(vector->z * vector->z)));
@@ -62,11 +62,11 @@ void	normalize(t_tuple *vector)
 	}
 }
 
-float	magnitude(t_tuple *vector)
+double	magnitude(t_tuple *vector)
 {
-	float	a;
-	float	b;
-	float	c;
+	double	a;
+	double	b;
+	double	c;
 
 	a = vector->x * vector->x;
 	b = vector->y * vector->y;
